# Slight keccak optimization