--- a/crypto/c_keccak.c
+++ b/crypto/c_keccak.c
@@ -70,19 +70,60 @@ void keccakf(uint64_t st[25], int rounds)
 		}
 
 		//  Chi
-		for (j = 0; j < 25; j += 5) {
-			bc[0] = st[j    ];
-			bc[1] = st[j + 1];
-			bc[2] = st[j + 2];
-			bc[3] = st[j + 3];
-			bc[4] = st[j + 4];
-			st[j    ] ^= (~bc[1]) & bc[2];
-			st[j + 1] ^= (~bc[2]) & bc[3];
-			st[j + 2] ^= (~bc[3]) & bc[4];
-			st[j + 3] ^= (~bc[4]) & bc[0];
-			st[j + 4] ^= (~bc[0]) & bc[1];
-		}
+		// unrolled loop, where only last iteration is different
+		j = 0;
+		bc[0] = st[j    ];
+		bc[1] = st[j + 1];
 
+		st[j    ] ^= (~st[j + 1]) & st[j + 2];
+		st[j + 1] ^= (~st[j + 2]) & st[j + 3];
+		st[j + 2] ^= (~st[j + 3]) & st[j + 4];
+		st[j + 3] ^= (~st[j + 4]) & bc[0];
+		st[j + 4] ^= (~bc[0]) & bc[1];
+
+		j = 5;
+		bc[0] = st[j    ];
+		bc[1] = st[j + 1];
+
+		st[j    ] ^= (~st[j + 1]) & st[j + 2];
+		st[j + 1] ^= (~st[j + 2]) & st[j + 3];
+		st[j + 2] ^= (~st[j + 3]) & st[j + 4];
+		st[j + 3] ^= (~st[j + 4]) & bc[0];
+		st[j + 4] ^= (~bc[0]) & bc[1];
+
+		j = 10;
+		bc[0] = st[j    ];
+		bc[1] = st[j + 1];
+
+		st[j    ] ^= (~st[j + 1]) & st[j + 2];
+		st[j + 1] ^= (~st[j + 2]) & st[j + 3];
+		st[j + 2] ^= (~st[j + 3]) & st[j + 4];
+		st[j + 3] ^= (~st[j + 4]) & bc[0];
+		st[j + 4] ^= (~bc[0]) & bc[1];
+
+		j = 15;
+		bc[0] = st[j    ];
+		bc[1] = st[j + 1];
+
+		st[j    ] ^= (~st[j + 1]) & st[j + 2];
+		st[j + 1] ^= (~st[j + 2]) & st[j + 3];
+		st[j + 2] ^= (~st[j + 3]) & st[j + 4];
+		st[j + 3] ^= (~st[j + 4]) & bc[0];
+		st[j + 4] ^= (~bc[0]) & bc[1];
+
+		j = 20;
+		bc[0] = st[j    ];
+		bc[1] = st[j + 1];
+		bc[2] = st[j + 2];
+		bc[3] = st[j + 3];
+		bc[4] = st[j + 4];
+
+		st[j    ] ^= (~bc[1]) & bc[2];
+		st[j + 1] ^= (~bc[2]) & bc[3];
+		st[j + 2] ^= (~bc[3]) & bc[4];
+		st[j + 3] ^= (~bc[4]) & bc[0];
+		st[j + 4] ^= (~bc[0]) & bc[1];
+		
 		//  Iota
 		st[0] ^= keccakf_rndc[round];
 	}