--- a/crypto/c_keccak.c
+++ b/crypto/c_keccak.c
@@ -24,18 +24,6 @@ const uint64_t keccakf_rndc[24] =
 	0x8000000000008080, 0x0000000080000001, 0x8000000080008008
 };
 
-const int keccakf_rotc[24] = 
-{
-	1,  3,  6,  10, 15, 21, 28, 36, 45, 55, 2,  14, 
-	27, 41, 56, 8,  25, 43, 62, 18, 39, 61, 20, 44
-};
-
-const int keccakf_piln[24] = 
-{
-	10, 7,  11, 17, 18, 3, 5,  16, 8,  21, 24, 4, 
-	15, 23, 19, 13, 12, 2, 20, 14, 22, 9,  6,  1 
-};
-
 // update the state with given number of rounds
 
 void keccakf(uint64_t st[25], int rounds)
@@ -63,11 +51,30 @@ void keccakf(uint64_t st[25], int rounds)
 
 		// Rho Pi
 		t = st[1];
-		for (i = 0; i < 24; ++i) {
-			bc[0] = st[keccakf_piln[i]];
-			st[keccakf_piln[i]] = ROTL64(t, keccakf_rotc[i]);
-			t = bc[0];
-		}
+		st[ 1] = ROTL64(st[ 6], 44);
+		st[ 6] = ROTL64(st[ 9], 20);
+		st[ 9] = ROTL64(st[22], 61);
+		st[22] = ROTL64(st[14], 39);
+		st[14] = ROTL64(st[20], 18);
+		st[20] = ROTL64(st[ 2], 62);
+		st[ 2] = ROTL64(st[12], 43);
+		st[12] = ROTL64(st[13], 25);
+		st[13] = ROTL64(st[19],  8);
+		st[19] = ROTL64(st[23], 56);
+		st[23] = ROTL64(st[15], 41);
+		st[15] = ROTL64(st[ 4], 27);
+		st[ 4] = ROTL64(st[24], 14);
+		st[24] = ROTL64(st[21],  2);
+		st[21] = ROTL64(st[ 8], 55);
+		st[ 8] = ROTL64(st[16], 45);
+		st[16] = ROTL64(st[ 5], 36);
+		st[ 5] = ROTL64(st[ 3], 28);
+		st[ 3] = ROTL64(st[18], 21);
+		st[18] = ROTL64(st[17], 15);
+		st[17] = ROTL64(st[11], 10);
+		st[11] = ROTL64(st[ 7],  6);
+		st[ 7] = ROTL64(st[10],  3);
+		st[10] = ROTL64(t, 1);
 
 		//  Chi
 		// unrolled loop, where only last iteration is different