I unrolled one of the loops in the `c_keccak.c` file and applied a slight optimization, noticing that one doesn't need to assign variables throughout the loop, just on the last iteration. It seems to have helped bump up the hash power a bit. Note that the highest H/s is higher for the PR than Dev. I repeated the test runs several times, always had an increase in highest H/s for PR than Dev (usually also, in the 2.5s, 60s, etc.).

**PR**
    
    HASHRATE REPORT                             
    | ID | 2.5s |  60s |  15m | ID | 2.5s |  60s |  15m |                                                                                                                        
    |  0 | 55.4 | 54.9 | (na) |  1 | 55.3 | 56.1 | (na) |                                                                                                                        
    -----------------------------------------------------                                                                                                                        
    Totals:   110.7 111.1 (na) H/s                                                                                                                                               
    Highest:  112.9 H/s                                                                                                                                                          
    RESULT REPORT                                                                                                                                                                
    Difficulty       : 29594                                                                                                                                                     
    Good results     : 5 / 5 (100.0 %)                                                                                                                                           
    Avg result time  : 19.4 sec                                                                                                                                                  
    Pool-side hashes : 102778                                                                                                                                                    
                                                                                                                                                                                 
    Top 10 best results found:                                                                                                                                                   
    |  0 |          2164579 |  1 |           413969 |                                                                                                                            
    |  2 |            48412 |  3 |            14799 |                                                                                                                            
    |  4 |             9432 |  5 |                0 |                                                                                                                            
    |  6 |                0 |  7 |                0 |                                                                                                                            
    |  8 |                0 |  9 |                0 |                                       
    
    Error details:                              
    Yay! No errors.

**Dev**

    HASHRATE REPORT                             
    | ID | 2.5s |  60s |  15m | ID | 2.5s |  60s |  15m |
    |  0 | 55.2 | 55.0 | (na) |  1 | 55.1 | 55.9 | (na) |
    -----------------------------------------------------
    Totals:   110.3 110.9 (na) H/s
    Highest:  112.7 H/s
    RESULT REPORT
    Difficulty       : 5210
    Good results     : 3 / 3 (100.0 %)
    Avg result time  : 36.0 sec
    Pool-side hashes : 9852
    
    Top 10 best results found:
    |  0 |            13160 |  1 |             5170 |
    |  2 |             3263 |  3 |                0 |
    |  4 |                0 |  5 |                0 |
    |  6 |                0 |  7 |                0 |
    |  8 |                0 |  9 |                0 |
    
    Error details:
    Yay! No errors.

It might need to be tested on other architectures. I only tested on my processor: Intel(R) Core(TM) i7-4600U CPU @ 2.10GHz. It has 4 cores and 4 MB cache.