@fireice-uk 2017-07-04
There are some further optimisations that can be done here, but I never bothered to make them as the whole keccak + tail hashes system benchmarks at around 1% of the runtime. 0.5% gain here sounds fairly high actually. We will try to reproduce it =)