@kathawala 2017-07-05
Thanks! And I've been looking at this for a while now, so it's fresh. If you tell me roughly what optimizations for this hash function you had in mind, I could try them out and make another PR.