@fireice-uk 2017-07-05
@kathawala See this code - https://github.com/redjack/nettle/blob/master/sha3-permute.c