@kathawala 2017-07-06
Thanks for the direction! I've made another optimization on the keccak function, which again resulted in a slight bump of H/s. Should I make a new pull request with both optimizations?