@fireice-uk 2017-07-06
@kathawala If you push any commits to your optimization branch they will be added here. Let me know when you are finished so we can verify the final result.