@kathawala 2017-07-08
Okay, I'm finished for now. I definitely won't work on any more optimizations for this routine in the upcoming week, so you can start verifying the results and see if this PR can be accepted. Thanks!