@psychocrypt 2017-07-09
@kathawala I will do some tests with you this weeks, thx for this PR