@psychocrypt 2017-07-12
I checked this PR and it is working but I think this is a job for the compiler. @fireice-uk  what do you think?