@fireice-uk 2017-07-12
@psychocrypt If we can keep the code readable, I always prefer to keep it readable. How does it actually benchmark (original vs manual vs pragma)?