@fireice-uk 2017-07-12
Also one problem, pragma unroll is specific to nvcc, having something that will unroll loop for both msvc and gcc might be difficult.