@psychocrypt 2017-07-12
> Also one problem, pragma unroll is specific to nvcc, having something that will unroll loop for both msvc and gcc might be difficult.

Ohh I thought all compiler knows this keyword even icc knows `#pragma unroll`

> How does it actually benchmark (original vs manual vs pragma)?

I can't see any difference between this branch and using `#pragma unroll` IMO the compiler knows that the loop can be unrolled because the iteration count is known at compile time. I am not sure if we need to force a unroll.

@fireice-uk I can also live with this PR even it looks more complicated. This provide use a speedup of 1.001, we should not spend to much time to improve this micro optimization.