@psychocrypt 2017-07-13
@fireice-uk  I approved this PR because gcc is not supporting `#pragma unroll` but I will give you the last word.