@fireice-uk 2017-07-13
I don't think it help significantly, but I don't see it hurting. I will merge it.